+++ packages/components/src/inputs/RadioButtonField/RadioButtonField.js
@@ -4,38 +4,44 @@ import classNames from 'classnames';
 import { Radio } from '../Radio';
 
 const RadioButtonField = ({
-  id, title, description, priceLabel, tabIndex, onChange, checked, className,
-}) => {
-  const [isChecked, setIsChecked] = useState(checked);
-  const handleChange = useCallback(() => {
-    setIsChecked(!isChecked);
-    onChange(checked);
-  }, [checked, isChecked]);
-
-  return (
-    <div
+  id,
+  title,
+  description,
+  priceLabel,
+  tabIndex,
+  onChange,
+  checked,
+  className,
+  name,
+  ...props
+}) => (
+  <label
     data-testid="mch-radio-button-field"
     tabIndex={tabIndex}
-    onClick={handleChange}
-    className={classNames('radioButtonField-container', className, { focus: isChecked })}>
-        <div className="radio-field-label">
-          <Radio
-            checked={isChecked}
-            id={id}
-            label={title}
-            onChange={onChange}
-            />
-          <div className="radio-label-container">
-            <p className="text-label-large">{title}</p>
-            <p className="text-small descriptor-text">{description}</p>
-          </div>
-        </div>
-        <div className="radio-price-container">
-          <p className="text-label-xlarge">{priceLabel}</p>
-        </div>
+    className={classNames('radioButtonField-container', className, {
+      focus: checked,
+    })}
+    htmlFor={id}
+  >
+    <div className="radio-field-label">
+      <Radio
+        checked={checked}
+        id={id}
+        label={title}
+        onChange={onChange}
+        name={name}
+        {...props}
+      />
+      <div className="radio-label-container">
+        <p className="text-label-large">{title}</p>
+        <p className="text-small descriptor-text">{description}</p>
+      </div>
     </div>
-  );
-};
+    <div className="radio-price-container">
+      <p className="text-label-xlarge">{priceLabel}</p>
+    </div>
+  </label>
+);
 
 RadioButtonField.propTypes = {
   id: PropTypes.string,
@@ -46,6 +52,7 @@ RadioButtonField.propTypes = {
   checked: PropTypes.bool,
   tabIndex: PropTypes.number,
   className: PropTypes.string,
+  name: PropTypes.string,
 };
 
 RadioButtonField.defaultProps = {